# config/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Node API",
    "description": "Node API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:5050",
  "basePath": "/api/v1/",
  "tags": [
    {
      "name": "Auth",
      "description": "API for authentication in the system"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "AuthHeader": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "description": "Login to the system",
        "parameters": [{
          "name": "auth",
          "in": "body",
          "description": "User auth details",
          "schema": {
            "type": "object",
            "required": ["username", "password"],
            "properties": {
              "username": {
                "type": "string"
              },
              "password": {
                "type": "string"
              }
            }
          }
        }],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "LOGGED_IN_SUCCESSFULLY"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "description": "Register to the system",
        "parameters": [{
          "name": "auth",
          "in": "body",
          "description": "User details",
          "schema": {
            "type": "object",
            "required": ["email"],
            "properties": {
              "email": {
                "type": "string"
              },
              "username": {
                "type": "string"
              }
            }
          }
        }],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "USER_REGISTERED"
          }
        }
      }
    }
  }
}